requirements.txt
pyyaml>=6.0
resampy>=0.2.2
numpy>=1.21
sounddevice
pyaudio
websockets
paho-mqtt>=1.6
